move `WasmAutoSave` to seperate file
