+++ bindings/wasm/src/account/account_builder.rs
@@ -21,27 +21,36 @@ use wasm_bindgen::__rt::WasmRefCell;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::future_to_promise;
+use crate::account::auto_save::WasmAutoSave;
 
+/// An [`Account`] builder for easy account configuration.
+///
+/// To reduce memory usage, accounts created from the same builder share the same `Storage`
+/// used to store identities, and the same {@link Client} used to publish identities to the Tangle.
+///
+/// The configuration on the other hand is cloned, and therefore unique for each built account.
+/// This means a builder can be reconfigured in-between account creations, without affecting
+/// the configuration of previously built accounts.
 #[wasm_bindgen(js_name = AccountBuilder)]
 pub struct WasmAccountBuilder(Rc<WasmRefCell<AccountBuilder>>);
 
 #[wasm_bindgen(js_class = AccountBuilder)]
 impl WasmAccountBuilder {
+  /// Creates a new `AccountBuilder`.
   #[wasm_bindgen(constructor)]
   pub fn new(options: Option<AccountBuilderOptions>) -> Result<WasmAccountBuilder> {
     let default_config: AccountConfig = AccountConfig::default();
     let mut builder = AccountBuilder::new();
 
     if let Some(o) = options {
-      let auto_save: AutoSave = match o.autoSave() {
-        Some(save) => save.0,
-        None => default_config.autosave,
-      };
-
       builder = builder
         .autopublish(o.autopublish().unwrap_or(default_config.autopublish))
         .milestone(o.milestone().unwrap_or(default_config.milestone))
-        .autosave(auto_save);
+        .autosave(
+          o.autoSave()
+            .map(|auto_save| auto_save.0)
+            .unwrap_or(default_config.autosave),
+        );
       //todo storage
       if let Some(mut config) = o.clientConfig() {
         let client: WasmClient = WasmClient::from_config(&mut config)?;
@@ -52,6 +61,8 @@ impl WasmAccountBuilder {
     Ok(Self(Rc::new(WasmRefCell::new(builder))))
   }
 
+  /// Loads an existing identity with the specified `did` using the current builder configuration.
+  /// The identity must exist in the configured `Storage`.
   #[wasm_bindgen(js_name = loadIdentity)]
   pub fn load_identity(&mut self, _did: WasmDID) -> Result<PromiseAccount> {
     todo!()
@@ -91,7 +102,7 @@ impl WasmAccountBuilder {
     }
 
     // Call the builder.
-    let builder = self.0.clone();
+    let builder: Rc<WasmRefCell<AccountBuilder>> = self.0.clone();
     let promise: Promise = future_to_promise(async move {
       builder
         .as_ref()
@@ -106,24 +117,7 @@ impl WasmAccountBuilder {
   }
 }
 
-#[wasm_bindgen(js_name = AutoSave)]
-pub struct WasmAutoSave(pub(crate) AutoSave);
 
-#[wasm_bindgen(js_class = AutoSave)]
-impl WasmAutoSave {
-  #[wasm_bindgen]
-  pub fn never() -> WasmAutoSave {
-    Self(AutoSave::Never)
-  }
-  #[wasm_bindgen]
-  pub fn every() -> WasmAutoSave {
-    Self(AutoSave::Every)
-  }
-  #[wasm_bindgen]
-  pub fn batch(number_of_actions: usize) -> WasmAutoSave {
-    Self(AutoSave::Batch(number_of_actions))
-  }
-}
 
 #[wasm_bindgen]
 extern "C" {
