+++ bindings/wasm/src/account/mod.rs
@@ -1,7 +1,8 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 mod account_builder;
 mod method_secret;
 mod update;
 mod wasm_account;
+mod auto_save;
